Cleaned up a bit of the messy stuff
