+++ voucher_service/pages/admin/dashboard.js
@@ -1,7 +1,9 @@
 import firebase from 'firebase';
 import React, { useEffect, useState } from 'react';
 import { useRouter } from 'next/router';
-import { firebaseConf } from '../../lib/config'
+import { firebaseConf } from '../../lib/config';
+import { getAdminUID } from '../../lib/firebaseUtil';
+import AdminDashboard from './AdminDashboard';
 
 const firebaseConfig = firebaseConf;
 
@@ -11,34 +13,9 @@ if (!firebase.apps.length) {
     firebase.app();
 }
 
-function userScreen() {
+export default function userScreen() {
     const [isSignedIn, setIsSignedIn] = useState(false);
-    const [isAdmin, setIsAdmin] = useState(false);
     const router = useRouter();
-
-    const logout = () => {
-        firebase.auth().signOut();
-        setIsSignedIn(false);
-        setIsAdmin(false);
-        router.replace('/login');
-    }
-
-    async function getAdminUID() {
-        const data = {key:null, value:null}
-    
-        const dbRef = firebase.database().ref();
-        await dbRef.child("Admin").child("Admin").get().then((snapshot) => {
-            if (snapshot.exists()) {
-                data.key = snapshot.key;
-                data.value = snapshot.val();
-            } else {
-                console.log("No data available");
-            }
-        }).catch((error) => {
-            console.log(error);
-        });
-        return data;
-    }
 
     useEffect(() => {
         const authObserver = firebase.auth().onAuthStateChanged(user => {
@@ -46,7 +23,7 @@ function userScreen() {
                 setIsSignedIn(!!user);
                 
                 // Check if the user is an admin
-                getAdminUID().then((data) => {
+                getAdminUID(firebase).then((data) => {
                     if (data.value == firebase.auth().currentUser.uid) {
                         // isAdmin
                     } else {
@@ -62,13 +39,13 @@ function userScreen() {
     }, []);
 
     if(isSignedIn) {
-        return (
-            <div>
-                <h1>Admin dashboard</h1>
-                <p>Welcome admin! You are now signed in!</p>
-                <button onClick={() => logout()}>Logout</button>
-            </div>
-        )
+        if(isSignedIn) {
+            return (
+                <AdminDashboard firebase={firebase}/>
+            )
+        } else {
+            router.replace('/login');
+        }
     }
 
     // loading screen goes here
@@ -78,5 +55,3 @@ function userScreen() {
         </div>
     )
 }
-
-export default userScreen;
